Editing of cmds and tests for the PSA app
