+++ psa/management/commands/clearpsa.py
@@ -2,7 +2,7 @@ from django.core.management.base import BaseCommand
 from django.core.management.color import no_style
 from django.db import connection
 
-from psa.models import Multimedia, Ecu
+from psa.models import Multimedia, Ecu, Corvet
 
 
 class Command(BaseCommand):
@@ -21,6 +21,12 @@ class Command(BaseCommand):
             dest='ecu',
             help='Clear Ecu table',
         )
+        parser.add_argument(
+            '--corvet',
+            action='store_true',
+            dest='corvet',
+            help='Delete all data in Corvet table',
+        )
 
     def handle(self, *args, **options):
 
@@ -31,7 +37,7 @@ class Command(BaseCommand):
             with connection.cursor() as cursor:
                 for sql in sequence_sql:
                     cursor.execute(sql)
-            self.stdout.write(self.style.SUCCESS("Suppression des données de la table Multimedia terminée!"))
+            self.stdout.write(self.style.WARNING("Suppression des données de la table Multimedia terminée!"))
         if options['ecu']:
             Ecu.objects.all().delete()
 
@@ -39,4 +45,12 @@ class Command(BaseCommand):
             with connection.cursor() as cursor:
                 for sql in sequence_sql:
                     cursor.execute(sql)
-            self.stdout.write(self.style.SUCCESS("Suppression des données de la table Ecu terminée!"))
+            self.stdout.write(self.style.WARNING("Suppression des données de la table Ecu terminée!"))
+        if options['corvet']:
+            Corvet.objects.all().delete()
+
+            sequence_sql = connection.ops.sequence_reset_sql(no_style(), [Corvet, ])
+            with connection.cursor() as cursor:
+                for sql in sequence_sql:
+                    cursor.execute(sql)
+            self.stdout.write(self.style.WARNING("Suppression des données de la table Corvet terminée!"))
